+++ tests/utils/sendMockSlackEvent.ts
@@ -1,4 +1,3 @@
-import { handler } from "@/functions/lambdas/slack-callback";
 import type { SlackCallbackRequest } from "@/types/SlackEventRequest";
 
 export const sendMockSlackEvent = async (body: SlackCallbackRequest) =>
@@ -11,49 +10,3 @@ export const sendMockSlackEvent = async (body: SlackCallbackRequest) =>
   })
     .then((res) => res.json())
     .catch((error) => console.error(error.stack));
-
-export const callWithMockSlackEvent = async (body: SlackCallbackRequest) =>
-  handler(
-    {
-      body: JSON.stringify(body),
-      headers: {},
-      isBase64Encoded: false,
-      rawPath: "",
-      rawQueryString: "",
-      requestContext: {
-        accountId: "",
-        apiId: "",
-        domainName: "",
-        domainPrefix: "",
-        http: {
-          method: "",
-          path: "",
-          protocol: "",
-          sourceIp: "",
-          userAgent: "",
-        },
-        requestId: "",
-        routeKey: "",
-        stage: "",
-        time: "",
-        timeEpoch: 0,
-      },
-      routeKey: "",
-      version: "",
-    },
-    {
-      awsRequestId: "",
-      callbackWaitsForEmptyEventLoop: false,
-      functionName: "",
-      functionVersion: "",
-      invokedFunctionArn: "",
-      logGroupName: "",
-      logStreamName: "",
-      memoryLimitInMB: "",
-      getRemainingTimeInMillis: () => 0,
-      done: () => {},
-      fail: () => {},
-      succeed: () => {},
-    },
-    () => {},
-  );
